+++ src/Components/Library/LibraryItemImage.jsx
